learnflow-app/code-runner/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
RestrictedPython==7.0
openai-agents